my-project/apps/services/src/sms_otp/requirements.txt
django-otp
twilio
flask
pymongo
python-dotenv